Stop output translation of `\n` into `\r\n`
Turning off OPOST stops output processing. By default the only output
processing feature is the translation of `\n` to `\r\n`.

This has the effect that when the `Enter` key is used the cursor moves
directly to the next line down, causing output to move diagonally down
to the right as keys are input and `Enter` is used.

A carriage return `\r` is needed to send the cursor to the start of the
line, so these have to be manually added in. `writefln()` already inserts
a `\n`.