Turn off miscellaneous flags
When BRKINT is turned on, a break condition will cause a SIGINT signal
to be sent to the program, like pressing `ctrl + c`.

INPCK enables parity checking, which doesn't seem to apply to modern
terminal emulators.

ISTRIP causes the 8th bit of each input byte to be stripped, meaning it
will set it to 0. This is likely already turned off.

CS8 is not a flag, it is a bit mask with multiple bits, which we set
unlike all of the flags that have been turned off. It sets the character
size (CS) to 8 bits per byte.