Add error handling
`perror()` looks at the global `errno` variable and prints a descriptive
error message for it. It also prints the string given to it before the
message, giving it context.

`tcsetattr()`, `tcgetattr()`, and `read()` all return -1 on failure, and
set the `errno` value.

In Cygwin when `read()` times out it returns -1 with an `errno` of
`EAGAIN` instead of just returning 0. To make deca work in Cygwin we
won't treat `EAGAIN` as an error.