Press `ctrl + q` to quit
In Kilos C source code this kind of function is done with a macro:
```
#define CTRL_KEY(k) ((k) & 0x1f)
```

D doesn't have a pre-processor. Instead the D way is to define a light
weight one line function, which the compiler optimizer will inline at
no loss of efficiency.

https://dlang.org/articles/pretod.html#macros

What we are doing here is a bitwise AND on the number 16 (1f in
hexadecimal) and a provided character. The character `q` is 113 in ASCII
encoding. The bitwise operation returns 17 which is the ASCII code for
`ctrl + q`. This bitwise operation works for all the control characters
because they have the ASCII encoding from 1-26. This mirrors what the
`ctrl` key does in the terminal: it strips the upper 3 bits from
whatever key you press in combination with `ctrl` and sends that.