Clear terminal screen
We are using `stdout.rawWrite()` to write 4 bytes out to the terminal.
These bytes make up an escape sequence.

`\x1b` is the escape character, 27 in decimal. Escape sequences start
with the escape character (27) followed by a `[` character. Escape
sequences instruct the terminal to do various formatting tasks, such as
coloring text, moving the cursor around, and clearing parts of the
screen.

Escape sequence commands take arguments, which come before the command.
We are using the `J` command to clear the screen. The argument is `2`
which says to clear the entire screen.

Calling `flush()` on `stdout` is required to send the output to the
terminal immediately, clearing the screen. Otherwise the screen could be
cleared at a later time e.g. program exit.

`stdout` needs to be fully qualified to avoid conflicting with
`core.stdc.stdio.stdout`.

For Deca we are mostly using VT100 escape sequences, which are supported
very widely by modern terminal emulators.

https://vt100.net/docs/vt100-ug/chapter3.html

If we wanted to support the maximum number of terminals out there we
could use the ncurses library.