Reposition the cursor to top left
This escape sequence is only 3 bytes long and uses the `H` command.

The `H` command takes two arguments: the row number and the column
number at which to position the cursor. On an 80x24 size terminal if you
want to position the cursor in the center of the screen, you could use
the command `<esc>[12;40H`. Multiple arguments are separated by a `;`
character.

The default arguments for `H` are 1. So leaving both arguments out will
position the cursor at the first row and first column, as if we had sent
the `<esc>[1;1H` command.

Rows and columns are numbered starting at 1, not 0, from the top left.