Draw the left gutter
`editorDrawRows()` handles drawing each row of the buffer of text being
edited. For now it draws a tilde (~) in each row, which means that row
is not part of the file and can't contain any text.

We don't know the size of the terminal yet, so we don't know how many
rows to draw. For now we just draw 24 rows.

After the rows are drawn we position the cursor to the top left again.

A note on `stdout.rawWrite()` It writes out a number of bytes to a file.
How many bytes is determined by the size of the buffer given to it to
write. In this case `leftGutter` is three character bytes, so three bytes
are written to `stdout`.