Scan stdout with the cursor for window size
`ioctl` is not guaranteed to be able to get the window size on all
systems. This change is the start of a fallback method of getting the
windows size.

We are positioning the cursor at the bottom-right of the screen. Future
changes will then query the position of the cursor, which will tell us
how many rows an columns there must be on the screen.

The escape sequence `\x1b[999C\x1b[999B` uses the `C` and `B` commands
to move the cursor 999 rows and columns. These commands are documented
not to go beyond the edge of the screen. This in theory should get us
to the maximum bounds of the screen.

https://vt100.net/docs/vt100-ug/chapter3.html#CUD

In this change we have `1 ||` at the front of the `if` condition
temporarily to test the fallback branch we are developing.

Kilo uses the C `write` function to write the escape sequence to stdout.
It is difficult to find and use the wrapped `write` function in D
without getting name conflicts. So I have used `rawWrite` instead. The
downside is that `write` returns the number of bytes written and
`rawWrite` returns `void`. So we can't test that the number of bytes
written is `== 12` and return `-1` on failure. The nearest D equivalent
I could find was to add a scope guard on failure.

At this point we are always returning `-1` (meaning an error occurred)
from `getWindowSize()` so we make a call to `editorReadKey()` so we can
observe the results of our escape sequences before the program calls
`die()`.

When running the program you should  see the cursor is positioned at the
bottom-right of the screen. However the cursor does not seem to move at
all, unless there is something is written to stdout. Then the output
appears at the bottom-right of the screen.