Report cursor position
We use the `n` command (Device status report) with an argument of `6` to
ask for the cursor position.

https://vt100.net/docs/vt100-ug/chapter3.html#DSR

We then read the reply from stdin and print the characters. In my case
the output is this at the bottom of the screen:
```
27
91 ('[')
53 ('5')
53 ('5')
59 (';')
49 ('1')
49 ('1')
52 ('4')
82 ('R')
```
The program immediately dies after.

This is an escape sequence. The escape character `27` followed by a `[`
then the actual response: `55;114R`. This is a cursor position report.

https://vt100.net/docs/vt100-ug/chapter3.html#CPR

It is reporting 55 lines and 114 columns.

The temporary call to `editorReadKey()` lets us observe debug output
before the screen gets cleared on exit.

The next change will parse this response.