Parse terminal 'report position' response
In my case the terminals response to the 'cursor position report' is
`<esc>[55;114R`: 55 lines and 114 columns.

In this change we read 32 bytes from stdin, slice the char buffer up to
the index of `R`, then print the response so we can see it.

When printing the parsed response we skip the first character `<esc>`
(27), otherwise the terminal would interpret it as an escape sequence
and not display it.

`parsedResponse[1..$]: '[55;114'` is written to stdout.