Minimize number of `rawWrite` calls by using a append buffer
There are about 5 writes to stdout each time the screen is refreshed.
Instead it would be better to do one big write to update the whole
screen at once. Otherwise there could be small unpredictable pauses
between writes, which would cause an annoying flicker effect.

In Kilo a custom C struct is written to create a dynamic string type
that supports appending. In D a dynamic array of type `char` is a
suitable alternative. Dynamic arrays have built in bounds checking,
avoiding buffer underflow and overflow problems. Also strings in D are
essentially arrays of characters.

The size of my terminal is 55 lines and 114 columns. When refreshing the
screen there will be 3 characters written to each line, excluding the
last line. 55 lines * 3 characters is 165 characters, so this is the
starting length of the `appendBuffer`.