Hide the cursor while refreshing the screen
There is another possible source of flicker effect while drawing the
screen. It is possible the cursor might be displayed in the middle of
the screen somewhere for a split second while the terminal is drawing to
the screen. To stop that from happening this change hides the cursor
before refreshing the screen, and shows it again immediately after.

The `h` and `l` commands, Get Mode and Reset Mode respectively, are used
to turn on and off various terminal modes. The VT100 User Guide does not
document the argument `?25`. It appeared in later VT models. So some
terminals might not support hiding/showing the cursor, in which case
they will ignore those escape sequences.

http://vt100.net/docs/vt100-ug/chapter3.html#SM
http://vt100.net/docs/vt100-ug/chapter3.html#RM
http://vt100.net/docs/vt100-ug/chapter3.html#S3.3.4
http://vt100.net/docs/vt510-rm/DECTCEM.html