Print welcome message
This shows the welcome message a third of the way down the screen. The
message prints the current version of Deca.

The versioning is done with Calendar versioning using full year,
short month, and short day: `YYYY.MM.DD`. I have gone with this instead
of semantic versioning because Deca is a stand alone complete
application. It is not intended to be a library or API.

From here git tags will be used to tag commits with the `decaVersion` of
that commit. As a new commit is added to `trunk` branch if the date of
the commit is at least one more day than the previous version then the
version should be updated to match the day of the commit and a git tag
should be added.