Move the cursor to the position in EditorConfig
We are using the `H` command to position the cursor. A locally scoped
function has been made to substitute the row and column values into the
escape sequence. By default the function will position the cursor at
the top left of the screen i.e. `x = 0` and `y = 0`.

When calling `cursorPosition` we add 1 to `E.cy` and `E.cx` to convert
from 0-indexed values to the 1-indexed values that the terminal uses.

Note that `appendbuffer` is type `char[]` while `cursorPosition` returns
a `string`. Strings are essentially an array of characters. There seems
to be no issue appending a `string` to a `char[]` without type casting.