Move cursor with WASD keys
I've changed the `switch` statements to `final` as these are statements
where an empty `default` case doesn't make sense. If a case range
statement needs to be used the `final` keyword will need to be removed.

Note that the program will crash if any key other than `wasd` or
`ctrl + q` is pressed.

There are some interesting bugs if the cursor is moved beyond the bounds
of the screen. The cursor might overflow to the opposite edge, or it may
stop at the edge. In both cases the underlying position can continue to
move further beyond the boundaries.