Move cursor with arrow keys
If `editorReadKey` reads an escape character `\x1b` it will then read
the next two bytes to see if it is an escape sequence. The buffer is
three bytes long because we will be handling longer escape sequences in
the future.

At the moment we are just aliasing the `WASD` keys. Future changes will
return special values for each arrow key that let us identify which
particular arrow key was pressed.