Enable `Page Up` and `Page Down` keys
The escape sequence for these keys is three bytes wide, which is why we
chose the length of `seq`.

We simulate the user pressing the up or down arrow enough times to move
to the bottom or top of the screen. Implementing it this way will help
when we implement scrolling.