Display text from open file
For now a hard coded string `Hello, world!` is a substitute for actual
file contents.

`Erow.size` needs to be of type `ulong` because that is the type that
`string.length` returns. The `size` field may not be required because
D strings/arrays know their length.

`editorDrawRows` will display the text of the file if the row it is
currently drawing contains a row from the file. In Kilo this row text
would be truncated if it went beyond the screen column width. That is
not done here as the D array append syntax `~=` does not have an easy
way of setting the number of characters to append. This may be a cause
of future bugs.