Display the first line of a file
The semantics of `File()` are the same as the `libc` `fopen` but wrapped
in a type that is much safer to use than directly manipulating a file
pointer.

This section is much simpler than Kilo because of the easier and safer
file and string abstractions. It is still not clear if `E.row.size` is
required as a D string knows its own length. The `Erow.chars` field type
is changed to `string` to fix type errors and because it is becoming
clear that D strings can do everything C character arrays can do and
more.

Because `Erow` represents an entire row of the editor we strip any
newline (and other white space) from the line read from the file.

The file to be opened is passed to Deca as a command line argument. If
no file is provided then `editorOpen()` will not be called and it will
start with a blank file.

This change only shows the first line. Future changes will show the
whole file.