Allocate memory for rows in EditorConfig
Normally structs are value types on the stack. In this case each Erow
will be an unknown size and there will be an unknown amount, potentially
a high number. Allocating on the heap seems more appropriate.

`EditorConfig` now has a dynamic array of `Erow` pointers. Since it is
known how many screen rows there are this is used as the initial size
of the array. It is made a dynamic array to ease appending new rows.

At this stage Deca will still only show the first line of a file.