Display all rows of a file on the screen
A plain text file containing the contents of the Text Editor topic from
Wikipedia has been created. It is used for testing Deca. The text is
formatted with line lengths up to 80 characters long. The text displays
well in my terminal screen.

There may be display errors when displaying text that has lines greater
in width that the terminal screen. While the terminal seems to wrap the
text, the cursor and top of the screen is positioned several lines down
the text file. Pressing the arrow or page down/up keys seems to cause a
repetition of the top few lines missing from the top of the screen.
These repeated lines can be seen by scrolling the terminal back up.

Future changes will cover vertical and horizontal scrolling. We will see
if these bugs persist after those changes.