Implement vertical scrolling
`EditorConfig.rowOffset` will keep track of what row of the file the
user is currently scrolled to.

`editorDrawRows` uses `rowOffset` to display the correct range of lines.

The value of `rowOffset` is set by checking if the cursor has moved
outside the visible window, and if it has, adjust `rowOffset` so that
the cursor is just inside the visible window. This is done in
`editorScroll()`, which is called when the screen is refreshed.

There might be a possible bug if the file being viewed contains tab
characters. The characters that the tabs take up are not being properly
erased when drawing to the screen. The issue will be fixed in future
changes.